frontend_server/requirements.txt
aiohttp==3.13.3
aiosignal>=1.3.2
async-generator==1.10
async-timeout==4.0.3
attrs==25.3.0
boto==2.49.0
botocore==1.38.0
certifi==2026.2.25
charset-normalizer==3.4.1
click==8.1.8
cycler==0.12.1
dj-database-url==3.1.2
Django==4.2.20
django-cors-headers==4.6.0
django-storages==1.14.4
exceptiongroup==1.2.2
frozenlist==1.5.0
gensim==4.4.0
gunicorn==25.1.0
h11==0.14.0
idna==3.10
importlib-metadata==8.6.1
jmespath==1.0.1
joblib>=1.4.2
kiwisolver==1.4.8
matplotlib==3.10.8
multidict==6.4.3
nltk==3.9.3
numpy==2.2.6
openai==2.26.0
outcome==1.3.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
Pillow==12.1.1
psycopg2-binary==2.9.11
qdrant-client>=1.9.0
pycparser==2.22
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2025.2
regex==2026.2.28
requests==2.32.5
s3transfer==0.11.3
scikit-learn==1.8.0
scikit-posthocs==0.12.0
scipy==1.17.1
seaborn==0.13.2
selenium==4.31.0
six==1.17.0
smart-open==7.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
statsmodels==0.14.6
threadpoolctl==3.6.0
tqdm==4.67.3
trio==0.29.0
trio-websocket==0.11.1
trueskill==0.4.5
typing-extensions==4.15.0
urllib3==2.4.0
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.0.0
django-allauth>=65.0.0
dj-rest-auth>=7.0.0
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.3.0
whitenoise==6.9.0
wsproto==1.2.0
yarl==1.20.0
yellowbrick==1.5
zipp==3.21.0
